refactor: CollapseLinks 경로 변경
