+++ src/Components/Nav/Item/LinkItem/index.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { NavLink } from 'react-router-dom';
-import CollapseLinks from '@Components/Nav/Item/LinkItem/CollapseLinks';
+import CollapseLinks from '@Components/Nav/Item/CollapseLinks';
 
 const LinkItem = props => {
     const item = props.item;
@@ -30,7 +30,12 @@ const LinkItem = props => {
                     {item.icon}
                     <span>{item.title}</span>
                 </div>
-                <CollapseLinks isOpened={expanded} setClose={setExpanded} items={item.children}></CollapseLinks>
+                <CollapseLinks
+                    id='nav'
+                    isOpened={expanded}
+                    setClose={setExpanded}
+                    items={item.children}
+                ></CollapseLinks>
             </li>
         );
     }
